Adding support for workflow_delete and workflow_sync and resturing directory structure
