+++ litmus-portal/cluster-agents/subscriber/pkg/requests/webhook.go
@@ -3,13 +3,13 @@ package requests
 import (
 	"encoding/json"
 	"errors"
+	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/utils"
 	"log"
 	"net/url"
 	"strings"
 	"time"
 
 	"github.com/gorilla/websocket"
-	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/events"
 	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/k8s"
 	"github.com/litmuschaos/litmus/litmus-portal/cluster-agents/subscriber/pkg/types"
 	"github.com/sirupsen/logrus"
@@ -125,7 +125,7 @@ func RequestProcessor(clusterData map[string]string, r types.RawData) error {
 			return errors.New("error performing cluster operation: " + err.Error())
 		}
 	} else if strings.Index("workflow_delete workflow_sync", strings.ToLower(r.Payload.Data.ClusterConnect.Action.RequestType)) >= 0 {
-		err := events.WorkflowRequest(clusterData, r.Payload.Data.ClusterConnect.Action.RequestType, r.Payload.Data.ClusterConnect.Action.ExternalData)
+		err := utils.WorkflowRequest(clusterData, r.Payload.Data.ClusterConnect.Action.RequestType, r.Payload.Data.ClusterConnect.Action.ExternalData)
 		if err != nil {
 			return errors.New("error performing events operation: " + err.Error())
 		}
